requirements.txt
altair==4.1.0
appdirs==1.4.3
argon2-cffi==20.1.0
asgiref==3.2.5
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
base58==2.0.1
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
boto3==1.14.38
botocore==1.17.38
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cvxopt==1.2.0
cvxpy==1.0.25 
cv2==4.2.0
pytesseract==0.3.7
dateparser==0.7.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
Django==3.0.3
docutils==0.15.2
docx2pdf==0.1.5
ecos==2.0.5
entrypoints==0.3
enum-compat==0.0.3
fastcache==1.1.0
feedparser==5.2.1
Flask==1.1.2
future==0.18.2
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
lxml==4.6.2
MarkupSafe==1.1.1
memory-profiler==0.58.0
mistune==0.8.4
mkl-service==2.3.0
mpmath==1.1.0
multiprocess==0.70.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
nose==1.3.7
notebook==6.2.0
numpy==1.19.5
optimize-images==1.3.6
ordered-set==3.1.1
osqp==0.6.1
packaging==20.9
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
patsy==0.5.1
pickleshare==0.7.5
piexif==1.1.3
Pillow==7.2.0
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.12.4
psutil==5.7.2
pyarrow==1.0.0
pycparser==2.20
pydeck==0.4.1
Pygments==2.8.0
pygooglenews==0.1.2
PyLaTeX==1.3.1
pylogit==0.2.2
pynput==1.7.1
Pyomo==5.6.8
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
PyQt5Designer==5.14.1
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.17.0
python-vlc==3.0.12117
pytz==2019.3
PyUtilib==5.7.3
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
redis==3.5.3
regex==2020.11.13
requests==2.24.0
s3transfer==0.3.3
scipy==1.3.1
scs==2.1.1.post2
Send2Trash==1.5.0
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.0
statsmodels==0.12.1
streamlit==0.64.0
sympy==1.5
terminado==0.9.2
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==6.1
tqdm==4.45.0
traitlets==4.3.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
validators==0.17.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.4.0